Feat | Carve Hull과 NavMesh Surface의 충돌 테스트 환경 구현

+++ Assets/02. Scripts/NavMeshResearch/MovableObstacle.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class MovableObstacle : MonoBehaviour
 {
     [SerializeField]
-    private int _movementSpeed;
+    private float _movementSpeed;
 
     private Vector3 _movementDirection;
     private bool _isMoving = false;
